[B] Ensure cover background image is correct size
The backgrounc cover image was tested on a Jolla C, but this has
slightly larger cover pages. When used on the Jolla 1, this
resulted in the background image appearing scaled up and cropped.
This fixes this, by using different images for Jolla 1 and
Jolla C devices (actually chosen by the width of the cover page).